Rename capture middleware
