+++ src/renderer/redux/app/thunks.ts
@@ -2,7 +2,7 @@ import { Action } from "redux";
 import { ThunkDispatch } from "redux-thunk";
 import { listDevices } from "../../services/imagingDevice/ImagingDevice";
 import * as rLogger from "../../services/rLogger/rLogger";
-import { changeDevice } from "../middleware/imagingDeviceMiddleware";
+import { changeDevice } from "../capture/imagingDeviceMiddleware";
 import { RootState } from "../store";
 import { setCurrentDevice, setDeviceList } from "./reducer";
