+++ src/renderer/redux/capture/middleware.ts
