+++ src/renderer/redux/store.ts
@@ -1,7 +1,7 @@
 import { applyMiddleware, combineReducers, createStore } from "redux";
 import thunk from "redux-thunk";
 import appReducer from "./app/reducer";
-import { createCaptureMiddleware } from "./middleware/imagingDeviceMiddleware";
+import { createCaptureMiddleware } from "./capture/middleware";
 import userPreferencesReducer from "./userPreferences/reducer";
 export type RootState = ReturnType<typeof rootReducer>;
